Remove functionalization path
