+++ thunder/core/update_aliases.py
@@ -1,10 +1,10 @@
 from functools import reduce, partial
 
-from thunder.core.functionalization import is_in_place_op, replace_args_with_alias_map
 import thunder.core.prims as prims
 from thunder.core.proxies import TensorProxy, variableify, unvariableify
-from thunder.core.symbol import BoundSymbolTag, has_tags
-from thunder.core.trace import from_trace, tracectx, TraceCtx as Trace, TraceProvenance
+from thunder.core.pytree import tree_flatten
+from thunder.core.symbol import BoundSymbol, BoundSymbolTag, has_tags
+from thunder.core.trace import from_trace, tracectx, TraceCtx as Trace, TraceProvenance, VariableInterface
 from thunder.torch import setitem_
 
 
@@ -75,8 +75,63 @@ def _add_output_to_setitem(trace: Trace) -> Trace:
     return trace
 
 
+def replace_args_with_alias_map(
+    computation_trace: Trace,
+    alias_tensor_indices: list[list[int]],
+) -> tuple[Trace, dict[VariableInterface, TensorProxy]]:
+    if not alias_tensor_indices:
+        return computation_trace, {}
+    bsyms: list[BoundSymbol] = []
+    flat_args, _ = tree_flatten((computation_trace.args, computation_trace.kwargs))
+    swap_map_for_aliases: dict[VariableInterface, TensorProxy] = {}
+    arg_to_optional_bsyms: dict[VariableInterface, BoundSymbol] = {}
+    for indices in alias_tensor_indices:
+        arg = flat_args[indices[0]]
+        for idx in filter(lambda idx: idx < len(flat_args), indices[1:]):
+            arg_to_replace = flat_args[idx]
+            reshaped_arg = arg
+            if arg_to_replace.shape != arg.shape:
+                with tracectx(computation_trace):
+                    reshaped_arg = prims.reshape.meta(arg, arg_to_replace.shape)
+                    arg_to_optional_bsyms[variableify(arg_to_replace)] = prims.reshape.bind(
+                        arg,
+                        arg_to_replace.shape,
+                        output=reshaped_arg,
+                    )
+            swap_map_for_aliases[variableify(arg_to_replace)] = reshaped_arg
+    appended_bsyms = {}
+    for bsym in computation_trace.bound_symbols:
+        for arg in filter(lambda p: isinstance(p, TensorProxy), bsym.flat_args):
+            reshape_bsym = arg_to_optional_bsyms.get(variableify(arg))
+            if reshape_bsym is not None:
+                if reshape_bsym not in appended_bsyms:
+                    bsyms.append(reshape_bsym)
+                    appended_bsyms[reshape_bsym] = arg
+        if replaced_args_map := {
+            x.name: swap_map_for_aliases[variableify(x)].name
+            for x in filter(lambda p: isinstance(p, TensorProxy), bsym.flat_args)
+            if variableify(x) in swap_map_for_aliases
+        }:
+            bsyms.append(bsym.from_bsym_swap_proxies(swap_map_for_aliases, skip_output=True))
+            if len(replaced_args_map) == 1:
+                bsyms[
+                    -1
+                ].header = f"[alias tensor args] `{list(replaced_args_map.keys())[0]}` is replaced by `{list(replaced_args_map.values())[0]}`"
+            else:
+                bsyms[
+                    -1
+                ].header = f"[alias tensor args] {list(replaced_args_map.keys())} are replaced by {list(replaced_args_map.values())}, respectively"
+        else:
+            bsyms.append(bsym)
+    no_implicit_alias_trace = from_trace(computation_trace)
+    no_implicit_alias_trace.bound_symbols = bsyms
+    str_map = {unvariableify(k).name: v.name for k, v in swap_map_for_aliases.items()}
+    no_implicit_alias_trace.set_provenance(TraceProvenance(f"Duplicate alias args using {str_map}"))
+    return no_implicit_alias_trace, swap_map_for_aliases
+
+
 def insert_alias_updates(computation_trace: Trace, alias_tensor_indices: list[list[int]]) -> Trace:
-    if not any(is_in_place_op(bsym) for bsym in computation_trace.bound_symbols):
+    if not any(_is_inplace_op(bsym) for bsym in computation_trace.bound_symbols):
         return computation_trace
 
     swap_map = dict()
